requirements.txt
discord.py>=2.3.0
treys>=0.1.8
aiosqlite>=0.19.0
python-dotenv>=1.0.0